requirements.txt
django
djangorestframework
numpy
pandas
matplotlib
